+++ test/runtests.jl
@@ -27,6 +27,7 @@ opequal(x, y) = x == y || x ≈ y
 
 include("test_bases.jl")
 include("test_tens.jl")
+include("test_tensor_products.jl")
 include("test_tens_isotropic.jl")
 include("test_tens_anisotropic.jl")
 include("test_tens_projection.jl")